Relocation of the query classes to core
