+++ java/clients/src/test/java/sleeper/clients/status/report/job/query/JobQueryTestBase.java
@@ -23,10 +23,10 @@ import sleeper.compaction.core.job.CompactionJob;
 import sleeper.compaction.core.job.CompactionJobStatusStore;
 import sleeper.compaction.core.job.CompactionJobStatusTestData;
 import sleeper.compaction.core.job.CompactionJobTestDataHelper;
-import sleeper.compaction.core.job.query.CompactionJobStatus;
 import sleeper.core.properties.instance.InstanceProperties;
 import sleeper.core.properties.table.TableProperties;
 import sleeper.core.properties.table.TableProperty;
+import sleeper.core.tracker.compaction.job.query.CompactionJobStatus;
 
 import java.time.Clock;
 import java.time.Instant;
