+++ java/compaction/compaction-core/src/test/java/sleeper/compaction/core/job/CompactionJobStatusInPeriodTest.java
@@ -17,9 +17,9 @@ package sleeper.compaction.core.job;
 
 import org.junit.jupiter.api.Test;
 
-import sleeper.compaction.core.job.query.CompactionJobStatus;
 import sleeper.core.record.process.RecordsProcessedSummary;
 import sleeper.core.record.process.RecordsProcessedSummaryTestHelper;
+import sleeper.core.tracker.compaction.job.query.CompactionJobStatus;
 
 import java.time.Instant;
