+++ java/core/src/main/java/sleeper/core/tracker/compaction/job/query/CompactionJobCreatedStatus.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package sleeper.compaction.core.job.query;
+package sleeper.core.tracker.compaction.job.query;
 
 import sleeper.core.record.process.status.ProcessStatusUpdate;
 import sleeper.core.statestore.AssignJobIdRequest;
