+++ java/system-test/system-test-dsl/src/main/java/sleeper/systemtest/dsl/reporting/CompactionReportsDriver.java
@@ -16,8 +16,8 @@
 
 package sleeper.systemtest.dsl.reporting;
 
-import sleeper.compaction.core.job.query.CompactionJobStatus;
 import sleeper.compaction.core.task.CompactionTaskStatus;
+import sleeper.core.tracker.compaction.job.query.CompactionJobStatus;
 
 import java.util.List;
